requirements.txt
RPi.GPIO
bottle